+++ narsha_front/src/pages/WriteScreen.tsx
